rTests/rMakev2/PublishProjectTests.cs
using rMakev2.DTOs;
using rMakev2.Models;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace rTests.rMakev2
{
    [TestClass]
    public class PublishProjectTests
    {
        
        [TestMethod]
        public void CreatePublishedProject_Ok() {
            //Arrange
            App testApp = new App(Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
            Data testData = new Data(testApp);
            Project testPtoject = new Project(testData);

            //Act
            PublishProject publishProject = new PublishProject(testPtoject);

            //Assert
            Assert.AreNotEqual(testPtoject.Id, publishProject.Id);
            Assert.AreEqual(DateTime.Now.Date, publishProject.PublicationDate);
        }

        [TestMethod]
        public void AddPublishedDocument() {
            //Arrange
            App testApp = new App(Guid.NewGuid().ToString(), Guid.NewGuid().ToString());
            Data testData = new Data(testApp);
            Project testPtoject = new Project(testData);

            Document Doc1 = testPtoject.Documents.First();
            PublishProject publishProject = new PublishProject(testPtoject);

            //Act
            PublishDocument publishDocument = publishProject.AddDocument(Doc1);

            //Assert
            Assert.IsTrue(publishDocument.Id == publishProject.Documents[1]);
        }
    }
}
